Change object deserialization strategy
Objects are now represented as classes with class objects in deserialization as
well. Several hacks added to aid in deserializing objects, see comments around
the code. "Class" proto message now can have a synthetic class object proto
within, if it's not contained in any external class file, which is the case
with objects. "class_object_present" boolean field is deprecated and will be
dropped, since its value is equivalent to the presence of "class_object" field