Change logic of serialization of primary constructors
"Callable primary_constructor" field will be dropped, because it requires
serializer to serialize the fact that a class has a primary constructor.
Default primary constructors for objects etc. shouldn't be serialized, so
there are three options: a class doesn't have a primary constructor, a class
has a default primary constructor, a class has a custom primary constructor.
This explains the change in descriptors.proto